Move helper functions to a separate script for importing
